services/telemetry/consumer/requirements.txt
confluent-kafka
SQLAlchemy
psycopg2-binary
python-dateutil